Modification page, improve permission controls, like system and delete publication

+++ back/controllers/publication.js
@@ -4,7 +4,6 @@ const Publication = require('../models/Publication');
 const User = require('../models/User');
 
 exports.createPublication = (req, res, next) => {
-    console.log(req.body);
     const publicationObject = req.body;
     delete publicationObject._id;
     const publication = new Publication({
@@ -12,9 +11,9 @@ exports.createPublication = (req, res, next) => {
         likes: 0,
         dislikes: 0,
         usersLiked: [],
-        usersDisliked: [],
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        usersDisliked: []
     });
+    req.file ? publication['imageUrl'] = `${req.protocol}://${req.get('host')}/images/${req.file.filename}` : "";
     publication.save()
         .then(() => res.status(201).json({ message: 'Ressource enregistré !' }))
         .catch(error => res.status(400).json({ error }));
@@ -29,14 +28,14 @@ exports.getOnePublication = (req, res, next) => {
 exports.modifyPublication = (req, res, next) => {
     Publication.findOne({ _id: req.params.id })
         .then(publication => {
-            const previousFilename = publication.imageUrl.split('/images/')[1];
+            const previousFilename = publication.imageUrl ? publication.imageUrl.split('/images/')[1] : undefined;
             const publicationObject = req.file? {
-                ...JSON.parse(req.body.publication),
+                ...req.body,
                 imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
             } : { ...req.body };
             Publication.updateOne({ _id: req.params.id }, { ...publicationObject, _id: req.params.id })
                 .then(() => {
-                    if (req.file) {
+                    if (previousFilename && req.file) {
                         fs.unlink(`images/${previousFilename}`, () => { });
                     }
                     res.status(200).json({message: 'Ressource modifiée !' })
@@ -49,12 +48,13 @@ exports.modifyPublication = (req, res, next) => {
 exports.deletePublication = (req, res, next) => {
     Publication.findOne({ _id: req.params.id })
         .then(publication => {
-            const filename = publication.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Publication.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: 'Ressource supprimée !' }))
-                    .catch(error => res.status(400).json({ error }));
-            });
+            if (publication.imageUrl) {
+                const filename = publication.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => { });
+            }
+            Publication.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Ressource supprimée !' }))
+                .catch(error => res.status(400).json({ error }));
         })
         .catch(error => res.status(500).json({ error }));
 };
@@ -70,31 +70,32 @@ exports.likePublication = (req, res, next) => {
     .then(() => {
         Publication.findOne({ _id: req.params.id })
         .then(publication => {
+            for (index in publication.usersLiked) {
+                if (req.body.userId === publication.usersLiked[index]) {
+                    publication.usersLiked.splice(index, 1);
+                    publication.likes -= 1;
+                }
+            }
+            for (index in publication.usersDisliked) {
+                if (req.body.userId === publication.usersDisliked[index]) {
+                    publication.usersDisliked.splice(index, 1);
+                    publication.dislikes -= 1;
+                }
+            }
             switch (req.body.like) {
                 case 1:
                     publication.usersLiked.push(req.body.userId);
                     publication.likes += 1;
                     break;
-                case 0:
-                    for (index in publication.usersLiked) {
-                        if (req.body.userId === publication.usersLiked[index]) {
-                            publication.usersLiked.splice(index, 1);
-                            publication.likes -= 1;
-                        }
-                    }
-                    for (index in publication.usersDisliked) {
-                        if (req.body.userId === publication.usersDisliked[index]) {
-                            publication.usersDisliked.splice(index, 1);
-                            publication.dislikes -= 1;
-                        }
-                    }
-                    break;
                 case -1:
                     publication.usersDisliked.push(req.body.userId);
                     publication.dislikes += 1;
                     break;
+                case 0:
+                    break;
                 default:
                     res.status(400).json({ message: 'Valeur de like invalide.' })
+                    return
             }
             Publication.updateOne({ _id: req.params.id }, { publication, _id: req.params.id, usersLiked: publication.usersLiked, usersDisliked: publication.usersDisliked, likes: publication.likes, dislikes: publication.dislikes})
                 .then(() => res.status(200).json({ message: 'Ressource mise à jour !' }))
